Attempt to bring sanity to MACHINE_ARCH.
The current method is to duplicate a bunch of logic between bootstrap,
bmake, and bsd.prefs.mk, with the possibility of overriding it with
mk.conf.  This just leads to pain, confusion, and multiple PRs.

Instead, move to a single source of truth - bootstrap.  This hardcodes
the correct value of MACHINE_ARCH into bmake, and all other settings of
MACHINE_ARCH are removed.  If MACHINE_ARCH is wrong for a particular
platform, then it should be fixed in bootstrap, nowhere else.

At the same time, try to sanitise the default setting of ABI, which
should fix issues seen on e.g. Linux where we didn't yet have the
default ABI setup in bsd.prefs.mk.  This can now also be removed.

Finally, LOWER_ARCH is unnecessary, so remove it too.  Update all
existing references to instead use MACHINE_ARCH.  Do the same with the
legacy SPARC_TARGET_ARCH variable.